+++ app/components/developers/specialties_component.rb
@@ -1,13 +1,14 @@
 module Developers
   class SpecialtiesComponent < ApplicationComponent
-    private attr_reader :specialties
+    private attr_reader :specialties, :force_show
 
-    def initialize(specialties)
+    def initialize(specialties, force_show: false)
       @specialties = specialties
+      @force_show = force_show
     end
 
     def render?
-      Feature.enabled?(:developer_specialties)
+      force_show || Feature.enabled?(:developer_specialty_querying)
     end
   end
 end